Update links and add spacing (#371)
* Update links and add spacing

Changed a few things, moved some parts, improved readability. 

Added links, mention Elm in Elm once more and add its contributor to the contributor list, as well as fixed some sentences and introduced Elm like spacing in order to improve the readability of the source. 

The line "In case of any questions about the project feel free to submit them in Issues with Q&A label" could create a new issue for the user - with q&a label already applied. Has to be done by a privileged person. 

Oh, and most important: Adds emoticons. :heart:

* Update README.md

I hope this helps

* Update README.md

Co-authored-by: Krzysztof Wende <krzysztof.wende@gmail.com>